+++ lib/mediator.dart
@@ -3,6 +3,7 @@ library flutter_mediator;
 export 'mediator/assert.dart';
 export 'mediator/controller.dart';
 export 'mediator/extension.dart';
+export 'mediator/global.dart';
 export 'mediator/host.dart';
 export 'mediator/multi_host.dart';
 export 'mediator/pub.dart';